backend/requirements.txt
# OKX Exchange SDK
python-okx>=5.3.0
websockets>=12.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Linting & Formatting
ruff>=0.2.0
